+++ nix-bindings-expr/build.rs
@@ -2,5 +2,5 @@ use nix_bindings_util::nix_version::emit_version_cfg;
 
 fn main() {
     let nix_version = pkg_config::probe_library("nix-expr-c").unwrap().version;
-    emit_version_cfg(&nix_version, &["2.26"]);
+    emit_version_cfg(&nix_version, &["2.26", "2.34.0pre"]);
 }